Fix: runtime error fix

+++ app/src/main/java/com/sas/companymanagement/ui/group/update/GroupUpdateFragment.kt
@@ -137,7 +137,7 @@ class GroupUpdateFragment :
 
         selectedArtistIdList.forEach { id ->
             binding.cgArtistUpdate.addView(Chip(context).apply {
-                artistUpdateViewModel.findArtistById(id.toInt())
+                artistUpdateViewModel.findArtistById(id)
                     .observe(viewLifecycleOwner) { artist ->
                         text = artist.artistName
                         isCloseIconVisible = true
